gui mostly implemented
GUI taking advantage of the browser frontend has been built. This necessitated large changes of the codebase but allows for much easier changes later.

Missing parts of the port are around crafting and inventory.

As a stop-gap measure, inventory is currently being displayed as a table within the player data section of the UI.

Crafting is not implemented, with the button being disabled by default.